remove pprint from Parameterized namespace
