+++ param/_utils.py
@@ -2,6 +2,8 @@ import inspect
 import functools
 import warnings
 
+from threading import get_ident
+
 
 class ParamDeprecationWarning(DeprecationWarning):
     """Param DeprecationWarning"""
@@ -64,3 +66,25 @@ def _deprecate_positional_args(func):
         return func(*args, **kwargs)
 
     return inner
+
+
+# Copy of Python 3.2 reprlib's recursive_repr but allowing extra arguments
+def _recursive_repr(fillvalue='...'):
+    'Decorator to make a repr function return fillvalue for a recursive call'
+
+    def decorating_function(user_function):
+        repr_running = set()
+
+        def wrapper(self, *args, **kwargs):
+            key = id(self), get_ident()
+            if key in repr_running:
+                return fillvalue
+            repr_running.add(key)
+            try:
+                result = user_function(self, *args, **kwargs)
+            finally:
+                repr_running.discard(key)
+            return result
+        return wrapper
+
+    return decorating_function
